api/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pandas>=2.1.0
openpyxl>=3.1.0
reportlab>=4.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
